examples/pydantic-ai/pyproject.toml
[project]
name = "pydantic-mcp-example"
version = "0.1.2"
description = "PydanticAI MCP example"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic-ai-slim[mcp,openai]>=0.3.5",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "ruff>=0.12.2",
]
